pyproject.toml
[project]
name = "linkedin-post"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = ["requests", "python-dotenv"]

[project.scripts]
linkedin-post = "linkedin.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[tool.hatch.build.targets.wheel]
packages = ["src/linkedin"]
